pyproject.toml
[project]
name = "agentic-dq-rule-onboarding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "fastmcp>=2.14.2",
    "google-adk>=1.14.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
